SSR: store processed documents on the Link instance to avoid memory leak
